0-print_listint.c

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -4,13 +4,13 @@
 /**
  * print_listint - print all the elements of a listint_t list
  * @h: head of linklist node
- * Return: the number nodes
+ * Return: the number of nodes
  *
  */
 
 size_t print_listint(const listint_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
